requirements.txt
appdirs==1.4.3
asgiref==3.4.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.2.8
django-extensions==3.1.3
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.2
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==0.6.2
numpy==1.21.2
openapi-codec==1.3.2
packaging==20.3
pandas==1.3.3
pep517==0.8.2
progress==1.5
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyparsing==2.4.6
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2021.3
requests==2.22.0
retrying==1.3.3
simplejson==3.17.5
six==1.14.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.3.0
